requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
coveralls==1.6.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.5
django-cors-headers==2.4.0
django-extensions==2.1.5
django-filter==2.1.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
django-simple-history==2.7.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docopt==0.6.2
drf-yasg==1.14.0
facebook-sdk==3.1.0
google-api-python-client==1.7.8
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.0.1
openapi-codec==1.3.2
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
python-twitter==3.5
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
ruamel.yaml==0.15.89
simplejson==3.16.0
six==1.12.0
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.1
validate-email==1.3
whitenoise==4.1.2
wrapt==1.11.1
